Add files via upload
The project aims to create an interactive 3D scenario that simulates an asteroid field. The scenario is
designed to allow users to navigate a spaceship through the field while experiencing realistic visual
and interactive effects.The project featured realistic lighting effects, as well as interactive features
that allow users to alter the scene, such as changing the color of spaceship using mouse interaction,
keyboard interaction for spaceship's movement and camera movement that follows the spaceship.